Update nabu jar (include bitswap client fixes)
Rename and simplify MetadataCachingStorage
