+++ src/peergos/server/Builder.java
@@ -226,10 +226,10 @@ public class Builder {
                 return s3;
             } else if (enableGC) {
                 TransactionalIpfs txns = new TransactionalIpfs(ipfs, transactions, authoriser, ipfs.id().join(), hasher);
-                return new MetadataCachingStorage(txns, meta, hasher);
+                return new MetadataWritingStorage(txns, meta);
             } else {
                 AuthedStorage target = new AuthedStorage(ipfs, authoriser, hasher);
-                return new MetadataCachingStorage(target, meta, hasher);
+                return new MetadataWritingStorage(target, meta);
             }
         } else {
             // In S3 mode of operation we require the ipfs id to be supplied as we don't have a local ipfs running
@@ -250,7 +250,7 @@ public class Builder {
                 return s3;
             } else {
                 FileContentAddressedStorage fileBacked = new FileContentAddressedStorage(blockstorePath(a), transactions, authoriser, hasher);
-                return new MetadataCachingStorage(fileBacked, meta, hasher);
+                return new MetadataWritingStorage(fileBacked, meta);
             }
         }
     }
